+++ familyhelper-ui-view/familyhelper-ui-view-pc/src/views/items/miscellaneous/fileEditor/txtSubEditor/TxtSubEditor.vue
@@ -26,34 +26,49 @@ export default {
       type: Boolean,
       default: true,
     },
+    contentedChanged: {
+      type: Boolean,
+      default: false,
+    },
   },
   watch: {
     blob() {
       this.syncBlob();
+    },
+    content() {
+      this.syncContentChanged();
     },
   },
   data() {
     return {
       content: '',
       busyFlag: false,
+      backupContent: '',
     };
   },
   methods: {
     syncBlob() {
       if (this.blob === null) {
         this.content = '';
+        this.backupContent = '';
       }
       this.busyFlag = true;
       this.blob.text()
         .then((text) => {
           this.content = text;
+          this.backupContent = text;
         })
         .finally(() => {
           this.busyFlag = false;
         });
     },
     contentToBlob() {
+      this.backupContent = this.content;
+      this.$emit('update:contentedChanged', false);
       return new Blob([this.content], { type: 'text/plain' });
+    },
+    syncContentChanged() {
+      this.$emit('update:contentedChanged', this.content !== this.backupContent);
     },
   },
   mounted() {